requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests==2.32.3
fpdf2==2.7.9